Update visitor.py, chainOfResponsibility.py, and 2 more files...

+++ 16. Market/market.py
@@ -1,5 +1,5 @@
 class Item:
-    def __init__(self, name: str, value: int):
+    def __init__(self, name: str, value: float):
         self.name = name
         self.value = value
